@@ -4,13 +4,13 @@ import java.text.SimpleDateFormat
 
 buildscript {
     ext {
-        version_gradle_tools = "3.6.3"
-        version_plugin_kotlin = "1.3.72"
-        enable_plugin_kotlin = false
+    version_gradle_tools = "8.2.0"
+    version_plugin_kotlin = "1.9.0"
+    enable_plugin_kotlin = false
 
-        version_compileSdk = 29
-        version_buildTools = "29.0.3"
-        version_minSdk = 17
+    version_compileSdk = 34
+    version_buildTools = "34.0.0"
+    version_minSdk = 23
 
         // https://developer.android.com/topic/libraries/support-library/
         version_library_appcompat = "28.0.0" //androidx
